Rename back the directory wrappers to astro
