+++ sql-cli/sql_cli/__main__.py
@@ -10,10 +10,10 @@ from rich import print as rprint
 from typer import Exit
 
 import sql_cli
+from sql_cli.astro.command import AstroCommand
+from sql_cli.astro.group import AstroGroup
 from sql_cli.constants import DEFAULT_AIRFLOW_HOME, DEFAULT_DAGS_FOLDER
 from sql_cli.exceptions import ConnectionFailed, DagCycle, EmptyDag, WorkflowFilesDirectoryNotFound
-from sql_cli.wrappers.command import AstroCommand
-from sql_cli.wrappers.group import AstroGroup
 
 if TYPE_CHECKING:
     from sql_cli.project import Project
