+++ sql-cli/sql_cli/astro/utils.py
